Moved accessibilityType to SSKeychainQuery.
This is to break the dependency of SSKeychainQuery on SSKeychain, so that the SSKeychainQuery can be used on its own.  It also allows those that use SSKeychainQuery on its own to specify accessibilityType on each query independently without having to change a static variable.

Also removed some redundant redefinitions of Security dictionary keys.